Added interests and generated target person
I have added two Json files one with a list of names which can be randomly cretaed by mixing a first and last name; the second will have a list of interests and their clues allowing for the person tobe randomly generated.

Next, I plan to make the first map area a small steet with a bar that the player will be able to enter.